refactor: import other components async in component doc  (#46491)
* refactor: ContributorAvatar component

* refactor: import other components async in component doc

* chore: fix lint

* docs: add suspense fallback

* docs: add suspense fallback

* docs: add suspense fallback
